Fixed coupon changing their code after redemption
`before_validation(on: :create)` erroneously runs with an already
persisted record again when it is added to a has_many relation.
In this case it is `redeemed_coupons << coupon`.
This caused the callback to run again and change the code which is not
supposed to happen. This is probably a bug in Rails and this change
works around it.