Repurpose archived attribute of events
The attribute will determine whether to show events in the archive in
the future.
Replace current method (that previously used the archived attribute)
with with_future_dates and add support for an offset that allows
to still get events with the last event being just in the past. This
allows box offices to still show event info after the event has already
strictly passed.