Fix certain received payments being falsely skipped
The hash of transactions was based on very limited information and
therefore often was not unique withing the same bank statement.
It now also takes the transaction details into account.
Also the most recent transaction is now determined using the date inside
the source rather than the creation date of the database record.